requirements.txt
aiogram==3.3.0
aiohttp~=3.9.0
python-dotenv==1.0.1
SQLAlchemy==2.0.34
greenlet==3.0.3
pydantic>=2.4.1,<2.6
alembic==1.13.1
aiosqlite==0.20.0
pydantic-settings==2.2.1 